usando interfaces

+++ ClassesObjetos/src/main/kotlin/Heranca/CardPayment.kt
@@ -3,4 +3,4 @@ package Heranca
 import java.math.BigDecimal
 import java.util.*
 
-class CardPayment(amount: BigDecimal, val number: String, val expiryDate: Date, val type: CardType) : Payment(amount)
+class CardPayment(amount: Float, val number: String, val expiryDate: Date, val type: CardType) : Payment(amount)
